// admcentro-web-main/src/interfaces/Iclaims.ts
// Generated by https://quicktype.io

export interface IClaimsResponse {
	results: number
	code: number
	status: string
	ok: boolean
	data: IClaim[]
}

export interface IClaim {
	details: Detail[]
	id: number
	PropertyId: number
	Property: {
		street: string
		number: string
		dept: string
		floor: string
	}
	state: State
	description: string
	createdAt: string
	updatedAt: string
	deletedAt: null
}

export interface Detail {
	comment?: string
	date: string
	id : string | number	
}

export enum State {
	Abierto = 'Abierto',
	Cerrado = 'Cerrado',
}